Add verb MLM features to the viz

+++ qfirst-scala/frame-browse/src/FeatureService.scala
@@ -60,10 +60,10 @@ object FeatureReq {
     type Out = Map[ArgumentId[Arg], Int]
   }
   case class ArgMLMDist[VerbType, Arg](verbType: VerbType, label: String) extends FeatureReq[VerbType, Arg] {
-    type Out = Map[ArgumentId[Arg], Map[String, Double]]
+    type Out = Map[ArgumentId[Arg], Map[String, Float]]
   }
   case class VerbMLMDist[VerbType, Arg](verbType: VerbType, label: String) extends FeatureReq[VerbType, Arg] {
-    type Out = Map[VerbId, Map[String, Double]]
+    type Out = Map[VerbId, Map[String, Float]]
   }
 
   // case class ArgMLMDist[Arg](pattern: String) extends ArgFeatureKey[Arg, Map[QuestionTemplate, Double]]
@@ -88,11 +88,11 @@ object FeatureReq {
       case ArgIndices(_) => implicitly[Encoder[List[(ArgumentId[Arg], Int)]]]
           .contramap[Map[ArgumentId[Arg], Int]](_.toList)
           .asInstanceOf[Encoder[req.Out]]
-      case ArgMLMDist(_, _) => implicitly[Encoder[List[(ArgumentId[Arg], List[(String, Double)])]]]
-          .contramap[Map[ArgumentId[Arg], Map[String, Double]]](_.iterator.map(p => p._1 -> p._2.toList).toList)
+      case ArgMLMDist(_, _) => implicitly[Encoder[List[(ArgumentId[Arg], List[(String, Float)])]]]
+          .contramap[Map[ArgumentId[Arg], Map[String, Float]]](_.iterator.map(p => p._1 -> p._2.toList).toList)
           .asInstanceOf[Encoder[req.Out]]
-      case VerbMLMDist(_, _) => implicitly[Encoder[List[(VerbId, List[(String, Double)])]]]
-          .contramap[Map[VerbId, Map[String, Double]]](_.iterator.map(p => p._1 -> p._2.toList).toList)
+      case VerbMLMDist(_, _) => implicitly[Encoder[List[(VerbId, List[(String, Float)])]]]
+          .contramap[Map[VerbId, Map[String, Float]]](_.iterator.map(p => p._1 -> p._2.toList).toList)
           .asInstanceOf[Encoder[req.Out]]
     }
   }
@@ -116,10 +116,10 @@ object FeatureReq {
       case ArgIndices(_) => implicitly[Decoder[List[(ArgumentId[Arg], Int)]]]
           .map(_.toMap)
           .asInstanceOf[Decoder[req.Out]]
-      case ArgMLMDist(_, _) => implicitly[Decoder[List[(ArgumentId[Arg], List[(String, Double)])]]]
+      case ArgMLMDist(_, _) => implicitly[Decoder[List[(ArgumentId[Arg], List[(String, Float)])]]]
           .map(_.iterator.map(p => p._1 -> p._2.toMap).toMap)
           .asInstanceOf[Decoder[req.Out]]
-      case VerbMLMDist(_, _) => implicitly[Decoder[List[(VerbId, List[(String, Double)])]]]
+      case VerbMLMDist(_, _) => implicitly[Decoder[List[(VerbId, List[(String, Float)])]]]
           .map(_.iterator.map(p => p._1 -> p._2.toMap).toMap)
           .asInstanceOf[Decoder[req.Out]]
     }
